Address PR review: shared billable helpers and hide empty tag settings
